pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmic_database"
version = "0.0.11"
authors = [
  { name="Ross Donnachie", email="code@radonn.co.za" },
]
description = "Python package simplifying access to SQL databases in use for COSMIC at the VLA."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
  "sqlalchemy",
  "pyaml",
]

[project.scripts]
cosmicdb_create_all_tables = "cosmic_database:engine.cli_create_all_tables"
cosmicdb_create_engine_url = "cosmic_database:engine.cli_create_engine_url"
